Fix: Removing a ride removes the ride from _world
Previously, when a ride was removed, it's information was not
removed from the world data structure. This caused a bug where
adding new ride after deleting a ride caused the voxels used by
the old ride to be rendered as if the were track pieces belonging
to the new ride.